Distributed_Event_Management_System/back-end/user-engagement/requirements.txt
boto3
flask
Werkzeug
flask-socketio
flask-cors
python-dotenv